feat: moved useFonts hook to own file
